// src/app/core/interfaces/GestionLlamada.interface.ts
export interface GestionLlamada {
    nombre: string;
    telefono: string;
    simpatizante: number;
    comentario: string;
    gestor: string;
    gestorId: number;
    distrito: number;
    claveUnica: string;
    titular: number;
    estatus: string;
    telefonoExtra: string;
    nombreExtra: string;
    direccion: string;
  }
  